requirements.txt
flask
flask-sqlalchemy
flask-migrate
flask-admin
python-dotenv
requests